refactor(jangar): split serving process status helpers
